+++ local_group_support/members.py
@@ -3,11 +3,9 @@ import json
 
 import pandas as pd
 
-from analysis_scripts.config.config import get_config
-from analysis_scripts.forms import get_forms
-from analysis_scripts.mattermost import post_to_channel
-from analysis_scripts.util import query, load_api_key
-from rebel_management_utilities import get_all_members
+from local_group_support.config.config import get_config
+from local_group_support.forms import get_forms
+from local_group_support.util import query
 
 FORMATION_DATE = datetime.date(2018, 4, 1)
 
@@ -88,9 +86,12 @@ def extract_data(member):
              'sign_up_channel': form} for form in forms]
 
 
-
 def get_member_stats(start_date):
-    members = get_all_members(api_key=load_api_key())
+    # members = get_all_members(api_key=load_api_key())
+
+    BACKUP_FILE_PATH = '/Users/pburghardt/Drive/Projekte/xr/rebel_backup/src/backups/backup_rebels_06-09-2020_10:35:07.json'
+    with open(BACKUP_FILE_PATH) as f:
+        members = json.load(f)['members']
 
     members_processed = []
